added order method to user prototype
