+++ README.md
@@ -1,0 +1,4 @@
+npm install;
+create database: jamazon-prime;
+
+npm run start:dev:seed;
